refactor(session): split storage.ts monolith into focused modules
2501-line storage.ts becomes a 72-line facade over storage-parse
(row/JSON parsing), storage-chains (chain CRUD), storage-history
(session/history loads), storage-compaction (apply transactions),
storage-sessions (list/CRUD), and storage-db (connection/recovery).
Complex parsers and apply transactions broken into cc<9 helpers.
Public import surface unchanged; session-persistence (148) and
chat-ipc (52) suites pass.
